backend/requirements.txt
asgiref==3.7.2
black==23.3.0
Django==4.2.6
django-colorfield==0.11.0
django-cors-headers==4.1.0
django-stubs==4.2.1
django-tinymce==3.6.0
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
djangorestframework-simplejwt==5.2.2
environs==9.5.0
flake8==6.0.0
isort==5.12.0
mypy==1.4.1
Pillow==10.1.0
psycopg2-binary==2.9.6
rest-social-auth==8.1.0
sqlparse==0.4.4
types-requests==2.31.0.10
